+++ src/main/java/com/github/antag99/retinazer/ComponentSet.java
@@ -14,6 +14,9 @@ import com.badlogic.gdx.utils.ObjectIntMap;
  */
 public final class ComponentSet {
 
+	/** Set of no components. */
+	public static final ComponentSet EMPTY = new ComponentSet();
+
 	private final Class<? extends Component>[] components;
 	private final ObjectIntMap<Class<? extends Component>> componentToId = new ObjectIntMap<>();
 
@@ -49,6 +52,41 @@ public final class ComponentSet {
 	 * @throws IndexOutOfBoundsException when the index does not belong to any component */
 	public Class<? extends Component> component(final int index) {
 		return components[index];
+	}
+
+	/** Check if this is a compatible subset of the other set.
+	 * A compatible subset is a subset where each component from the subset has the same index in both sets. */
+	public boolean isSubsetOf(ComponentSet other) {
+		final Class<? extends Component>[] myComponents = this.components;
+		final Class<? extends Component>[] otherComponents = other.components;
+		final int mySize = myComponents.length;
+		if (mySize > otherComponents.length) {
+			return false;
+		}
+
+		for (int i = 0; i < mySize; i++) {
+			if (!myComponents[i].equals(otherComponents[i])) {
+				return false;
+			}
+		}
+		return true;
+	}
+
+	/** Create a new {@link FamilySpec} which contains all entities. */
+	public final FamilySpec family() {
+		return new FamilySpec(this);
+	}
+
+	/** Create a new {@link FamilySpec} which requires given components. */
+	@SafeVarargs
+	public final FamilySpec familyWith(Class<? extends Component>... components) {
+		return new FamilySpec(this, true, components);
+	}
+
+	/** Create a new {@link FamilySpec} which excludes given components. */
+	@SafeVarargs
+	public final FamilySpec familyWithout(Class<? extends Component>... components) {
+		return new FamilySpec(this, false, components);
 	}
 
 	Mapper<?>[] buildComponentMappers(Engine engine) {