+++ src/main/java/com/github/antag99/retinazer/EntitySetView.java
@@ -3,13 +3,7 @@ package com.github.antag99.retinazer;
 import com.badlogic.gdx.utils.IntArray;
 import com.github.antag99.retinazer.util.Mask;
 
-public final class EntitySetView {
-
-    private final EntitySet content;
-
-    EntitySetView(EntitySet content) {
-        this.content = content;
-    }
+public interface EntitySetView {
 
     /**
      * Checks if this set contains the given entity.
@@ -18,9 +12,7 @@ public final class EntitySetView {
      *            the entity to check for.
      * @return whether this set contains the given entity.
      */
-    public boolean contains(int entity) {
-        return content.contains(entity);
-    }
+    boolean contains(int entity);
 
     /**
      * Returns the entities contained in this entity set.
@@ -28,9 +20,7 @@ public final class EntitySetView {
      *
      * @return the entities contained in this set.
      */
-    public Mask getMask() {
-        return content.getMask();
-    }
+    Mask getMask();
 
     /**
      * Returns an array containing the indices of all entities in this set.
@@ -39,30 +29,9 @@ public final class EntitySetView {
      *
      * @return the indices of all entities in this set.
      */
-    public IntArray getIndices() {
-        return content.getIndices();
-    }
+    IntArray getIndices();
 
-    public int size() {
-        return content.size();
-    }
+    int size();
 
-    public boolean isEmpty() {
-        return content.isEmpty();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return obj instanceof EntitySetView && ((EntitySetView) obj).content.equals(content);
-    }
-
-    @Override
-    public int hashCode() {
-        return ~content.hashCode();
-    }
-
-    @Override
-    public String toString() {
-        return "View "+content.toString();
-    }
+    boolean isEmpty();
 }