+++ src/main/java/com/github/antag99/retinazer/FamilyManager.java
@@ -1,15 +1,14 @@
 package com.github.antag99.retinazer;
 
 import com.badlogic.gdx.utils.ObjectIntMap;
-import com.badlogic.gdx.utils.ObjectSet;
 import com.github.antag99.retinazer.util.Bag;
 import com.github.antag99.retinazer.util.Mask;
 
 final class FamilyManager {
 
     private static final class Key {
-        ObjectSet<Class<? extends Component>> components = null;
-        ObjectSet<Class<? extends Component>> excludedComponents = null;
+        Mask requiredComponents = null;
+        Mask excludedComponents = null;
 
         @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
         @Override
@@ -19,72 +18,56 @@ final class FamilyManager {
             // No need for a type check; this class is only used internally
             Key key = (Key) obj;
             return key.excludedComponents.equals(excludedComponents) &&
-                    key.components.equals(components);
+                    key.requiredComponents.equals(requiredComponents);
         }
 
         @Override
         public int hashCode() {
             // Excluded components are rarer than required components; prioritize
             // the components hashCode over excluded components hashCode.
-            return 31 * excludedComponents.hashCode() + components.hashCode();
+            return 31 * excludedComponents.hashCode() + requiredComponents.hashCode();
         }
     }
 
     private final ObjectIntMap<Key> familyIndices = new ObjectIntMap<>();
     private final Bag<Family> families = new Bag<>();
     private final Engine engine;
-    private EntitySetView entities;
 
-    public FamilyManager(Engine engine) {
+    FamilyManager(Engine engine) {
         this.engine = engine;
-    }
-
-    public EntitySetView getEntities() {
-        if (entities == null) {
-            entities = getFamily(new FamilyConfig()).getEntities();
-        }
-        return entities;
     }
 
     private final Key getFamily_lookup = new Key();
 
     /** @return existing or new family conforming to the given configuration */
-    public Family getFamily(FamilyConfig config) {
-        final Key lookup = this.getFamily_lookup;
-        lookup.components = config.components;
-        lookup.excludedComponents = config.excludedComponents;
-        final int index = familyIndices.get(lookup, familyIndices.size);
+    public Family getFamily(FamilySpec spec) {
+        assert spec.domain.isSubsetOf(engine.componentDomain);
+        final ObjectIntMap<Key> familyIndices = this.familyIndices;
+        final int index;
+        {
+            final Key lookup = this.getFamily_lookup;
+            lookup.requiredComponents = spec.requiredComponents;
+            lookup.excludedComponents = spec.excludedComponents;
+            index = familyIndices.get(lookup, familyIndices.size);
+        }
 
         if (index == familyIndices.size) {
-            int i;
-            int[] components = new int[config.components.size];
-            int[] excludedComponents = new int[config.excludedComponents.size];
-
-            i = 0;
-            for (Class<? extends Component> componentType : config.components)
-                components[i++] = engine.componentSet.index(componentType);
-
-            i = 0;
-            for (Class<? extends Component> componentType : config.excludedComponents)
-                excludedComponents[i++] = engine.componentSet.index(componentType);
-
-            Family family = new Family(components, excludedComponents, index);
+            final Family family = new Family(spec.requiredComponents, spec.excludedComponents, index);
             Key key = new Key();
-            key.components = config.components;
-            key.excludedComponents = config.excludedComponents;
+            key.requiredComponents = spec.requiredComponents;
+            key.excludedComponents = spec.excludedComponents;
             familyIndices.put(key, index);
             families.set(index, family);
 
             // Find matching entities, and add them to the new family set.
             final Mapper<?>[] mappers = engine.componentMappers;
             Mask matchedEntities = new Mask().set(engine.entities);
-
-            for (int component : components) {
-                matchedEntities.and(mappers[component].componentsMask);
+            for (int componentI = spec.requiredComponents.nextSetBit(0); componentI != -1; componentI = spec.requiredComponents.nextSetBit(componentI + 1)) {
+                matchedEntities.and(mappers[componentI].componentsMask);
             }
 
-            for (int excludedComponent : excludedComponents) {
-                matchedEntities.andNot(mappers[excludedComponent].componentsMask);
+            for (int componentI = spec.excludedComponents.nextSetBit(0); componentI != -1; componentI = spec.excludedComponents.nextSetBit(componentI + 1)) {
+                matchedEntities.andNot(mappers[componentI].componentsMask);
             }
 
             // No notifications to dispatch here
@@ -103,65 +86,63 @@ final class FamilyManager {
      * to/from family sets.
      */
     void updateFamilyMembership() {
-        final Mapper<?>[] mappers = engine.componentMappers;
+        final Engine engine = this.engine;
+        final Bag<Family> families = this.families;
+        final int familyCount = familyIndices.size;
 
-        final Mask tmpMask = this.updateFamilyMembership_tmpMask;
-        @SuppressWarnings("UnnecessaryLocalVariable")
-        final Mask tmpMatchedEntities = this.updateFamilyMembership_tmpMatchedEntities;
+        {
+            final Mapper<?>[] mappers = engine.componentMappers;
+            final Mask tmpMask = this.updateFamilyMembership_tmpMask;
+            final Mask matchedEntities = this.updateFamilyMembership_tmpMatchedEntities;
+            for (int i = 0; i < familyCount; i++) {
+                Family family = families.get(i);
+                assert family != null;
+                EntitySet entities = family.entities;
 
-        for (int i = 0, n = familyIndices.size; i < n; i++) {
-            Family family = families.get(i);
-            assert family != null;
-            EntitySet entities = family.entities;
+                matchedEntities.set(engine.entities);
+                matchedEntities.andNot(engine.remove);
 
-            Mask matchedEntities = tmpMatchedEntities.set(engine.entities);
-            matchedEntities.andNot(engine.remove);
+                for (int componentI = family.requiredComponents.nextSetBit(0); componentI != -1; componentI = family.requiredComponents.nextSetBit(componentI + 1)) {
+                    final Mapper<?> mapper = mappers[componentI];
+                    tmpMask.set(mapper.componentsMask);
+                    tmpMask.andNot(mapper.removeMask);
+                    matchedEntities.and(tmpMask);
+                }
 
-            int[] components = family.components;
-            for (int component : components) {
-                Mapper<?> mapper = mappers[component];
-                tmpMask.set(mapper.componentsMask);
-                tmpMask.andNot(mapper.removeMask);
-                matchedEntities.and(tmpMask);
+                for (int componentI = family.excludedComponents.nextSetBit(0); componentI != -1; componentI = family.excludedComponents.nextSetBit(componentI + 1)) {
+                    final Mapper<?> mapper = mappers[componentI];
+                    tmpMask.set(mapper.componentsMask);
+                    tmpMask.andNot(mapper.removeMask);
+                    matchedEntities.andNot(tmpMask);
+                }
+
+                family.insertEntities.set(matchedEntities);
+                family.insertEntities.andNot(entities.getMask());
+                entities.addEntities(family.insertEntities);
+
+                family.removeEntities.set(entities.getMask());
+                family.removeEntities.andNot(matchedEntities);
+                entities.removeEntities(family.removeEntities);
             }
-
-            int[] excludedComponents = family.excludedComponents;
-            for (int excludedComponent : excludedComponents) {
-                Mapper<?> mapper = mappers[excludedComponent];
-                tmpMask.set(mapper.componentsMask);
-                tmpMask.andNot(mapper.removeMask);
-                matchedEntities.andNot(tmpMask);
-            }
-
-            family.insertEntities.set(matchedEntities);
-            family.insertEntities.andNot(entities.getMask());
-            entities.addEntities(family.insertEntities);
-
-            family.removeEntities.set(entities.getMask());
-            family.removeEntities.andNot(matchedEntities);
-            entities.removeEntities(family.removeEntities);
         }
 
-        final EntitySet argument = this.updateFamilyMembership_argument;
+        {
+            final EntitySet argument = this.updateFamilyMembership_argument;
+            for (int i = 0; i < familyCount; i++) {
+                Family family = families.get(i);
+                assert family != null;
 
-        for (int i = 0, n = familyIndices.size; i < n; i++) {
-            Family family = families.get(i);
-            assert family != null;
-
-            if (!family.insertEntities.isEmpty()) {
-                argument.addEntities(family.insertEntities);
-                for (EntityListener listener : family.listeners) {
-                    listener.inserted(argument.view());
+                if (!family.insertEntities.isEmpty()) {
+                    argument.addEntities(family.insertEntities);
+                    family.onInserted(argument);
+                    argument.clear();
                 }
-                argument.clear();
-            }
 
-            if (!family.removeEntities.isEmpty()) {
-                argument.addEntities(family.removeEntities);
-                for (EntityListener listener : family.listeners) {
-                    listener.removed(argument.view());
+                if (!family.removeEntities.isEmpty()) {
+                    argument.addEntities(family.removeEntities);
+                    family.onRemoved(argument);
+                    argument.clear();
                 }
-                argument.clear();
             }
         }
     }