+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Changelog
 
 # Version 0.2.3-SNAPSHOT
-- It is now required to supply all components at the `Engine` creation
+- It is now required to supply all components at the `Engine` creation through `ComponentSet`
 - `EngineConfig` has been deleted
 - `Engine.getSystem` has been replaced with generic `getService`
+- Renamed `FamilyConf` to `FamilySpec` and made it immutable
 - `ensureCapacity(int)` for all `Bag` implementations
 - Fix: `Bag.set(int)` no longer treats the `0` value specially